requirements.txt
Flask
Flask-Cors
python-dotenv
openai
